Updated method for handling arb size items; added dslv.py tool
Yet again, I have changed the algorithm that calculates the
new position of the floating (dragging) list item. This time,
the fundamentals are there for smooth handling of arbitrary
height list items.

Also added was a debugging tool, named dslv.py. The DSLV inner
class DragScroller now has the ability to dump the state of
the DSLV every frame of a drag scroll (via the new inner class:
StateTracker). State is dumped to an xml file; dslv.py parses
that file to build an interactive replay of the last drag
scroll. Frame-by-frame replays makes debugging a little less
nightmarish.