Improve long-press support
When starting a drag, set the dragged item to be not pressed before
taking a snapshot of its drawing cache.
This fixes the problem of the dragged item sometimes being highlighted
or invisible, depending on when you started the drag.

Add support for only starting a drag with a long-press. This allows you
to use the whole item layout as the drag handle but still let scrolling
work. Apologies for the name of the new attribute enumerator: using
"drag" makes aapt segfault because it has the same name as an existing
ID.

Start differentiating horizontal and vertical drags. This means that
when you have a remove handler but no drag handler, a vertical drag
still scrolls the list, but a horizontal drag can remove the item. This,
again, is especially useful when the whole item layout is the drag
handle, and makes the list work like Android's recent apps list. More
work is needed to distinguish horizontal and vertical drags when both
are enabled.

Add a demo for long-press-to-sort.

Add some text in the readme about (a) the new sort_mode attribute, and
(b) complications of having the drag handle be the whole list item.