Make item layout current for onDragFloatView
User updates of the floating View performed in the
onDragFloatView callback may rely on the positions of
the items in the ListView (e.g. when drags should be
limited to a subset of the list). This commit makes
efforts to guarantee that the item positions are
updated and in the most recently-requested state
before the callback is invoked.