Add DragSortCursorAdapter helper classes
These classes take care of virtualizing the reordering of
data coming from a Cursor. They act just like the
CursorAdapter, ResourceCursorAdapter, and SimpleCursorAdapter
classes from the android library.

Also, since it is common to implement the drag-sort Listeners
as part of a custom ListAdapter, DSLV now recognizes if
the ListAdapter given to it implements a Listener interface.
If so, the provided ListAdapter replaces the current
Listeners.