Optimize item shuffling underneath floating View
Bad coding was causing the list items to be remeasured very
often during a drag-sort. No more. Note the new class
DragSortItemView; this is a very simple ViewGroup whose purpose
is to display a single item. It takes care to avoid calls to
measure(int, int) on its child View when they are unnecessary
(like during a drag-sort, when the height of the DragSortItemView
changes frequently, but the contained View does not).

Another optimization was added that caches the heights of list
items that are offscreen. This should, in principle, give
better performance during drag-scrolls, especially for complex
list item Views.