Change from item wrapper to gaps layout technique
Remove DragSortListItem class and accompanying method
for creating gaps between list items. A summary:

Prior to this commit, each list item was wrapped in a
lightweight ViewGroup called DragSortListItem. DSLV
would modify the height of the wrapper (leaving the
height of the wrapped layout unchanged) to create blank
spaces between list items so as to indicate the new
position into which the dragged item might be dropped.

The new behavior does away with the DragSortListItem
wrapper and, instead, shifts the list items away from
each other to create the drop/shuffle gaps.
A few advantages over the old method come to mind:

- Removes unnecessary added layer in the View hierarchy.
- Item shuffling avoids layout passes (unless scrolling),
simply calling `invalidate()`.
- Code is more elegant and maintainable (previous hacks
to improve shuffling and scrolling smoothness magically
disappear).